Split up generic `AddSerializer`
Into three distinct ones for each case.
