+++ app/services/create_featured_tag_service.rb
@@ -26,6 +26,6 @@ class CreateFeaturedTagService < BaseService
   private
 
   def build_json(featured_tag)
-    Oj.dump(serialize_payload(featured_tag, ActivityPub::AddSerializer, signer: @account))
+    Oj.dump(serialize_payload(featured_tag, ActivityPub::AddHashtagSerializer, signer: @account))
   end
 end
